+++ database/migrations/2021_09_18_163903_ground_and_delivery_prenotes.php
@@ -1,0 +1,34 @@
+<?php
+
+use App\Services\DependencyService;
+use Carbon\Carbon;
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\DB;
+
+class GroundAndDeliveryPrenotes extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        DB::table('controller_positions')
+            ->where('callsign', 'LIKE', '%_GND')
+            ->orWhere('callsign', 'LIKE', '%_DEL')
+            ->update(['sends_prenotes' => true, 'updated_at' => Carbon::now()]);
+        
+        DependencyService::touchDependencyByKey('DEPENDENCY_CONTROLLER_POSITIONS_V2');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        //
+    }
+}